Move host config definition to separate file
