+++ pkg/controller/baremetalhost/baremetalhost_controller.go
@@ -27,7 +27,6 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
@@ -153,64 +152,6 @@ type reconcileInfo struct {
 // match the provisioner.EventPublisher interface
 func (info *reconcileInfo) publishEvent(reason, message string) {
 	info.events = append(info.events, info.host.NewEvent(reason, message))
-}
-
-// hostConfigData is an implementation of host configuration data interface.
-// Object is able to retrive data from secrets referenced in a host spec
-type hostConfigData struct {
-	host   *metal3v1alpha1.BareMetalHost
-	log    logr.Logger
-	client client.Client
-}
-
-// Generic method for data extraction from a Secret. Function uses dataKey
-// parameter to detirmine which data to return in case secret contins multiple
-// keys
-func (hcd *hostConfigData) getSecretData(name, namespace, dataKey string) (string, error) {
-	secret := &corev1.Secret{}
-	key := types.NamespacedName{
-		Name:      name,
-		Namespace: namespace,
-	}
-	if err := hcd.client.Get(context.TODO(), key, secret); err != nil {
-		errMsg := fmt.Sprintf("failed to fetch user data from secret %s defined in namespace %s", name, namespace)
-		return "", errors.Wrap(err, errMsg)
-	}
-
-	data, ok := secret.Data[dataKey]
-	if !ok {
-		hostConfigDataError.WithLabelValues(dataKey).Inc()
-		return "", NoDataInSecretError{secret: name, key: dataKey}
-	}
-
-	return string(data), nil
-}
-
-// UserData get Operating System configuration data
-func (hcd *hostConfigData) UserData() (string, error) {
-	if hcd.host.Spec.UserData == nil {
-		hcd.log.Info("UserData is not set return empty string")
-		return "", nil
-	}
-	return hcd.getSecretData(
-		hcd.host.Spec.UserData.Name,
-		hcd.host.Spec.UserData.Namespace,
-		"userData",
-	)
-
-}
-
-// NetworkData get network configuration
-func (hcd *hostConfigData) NetworkData() (string, error) {
-	if hcd.host.Spec.NetworkData == nil {
-		hcd.log.Info("NetworkData is not set returning epmty(nil) data")
-		return "", nil
-	}
-	return hcd.getSecretData(
-		hcd.host.Spec.NetworkData.Name,
-		hcd.host.Spec.NetworkData.Namespace,
-		"networkData",
-	)
 }
 
 // Reconcile reads that state of the cluster for a BareMetalHost
